Extract out closely related code as classes
