+++ src/main/java/TaskList.java
@@ -2,23 +2,21 @@ import java.util.ArrayList;
 
 public class TaskList {
     ArrayList<Task> list;
+    Ui ui;
     private static final String line =
             "____________________________________________________________";
 
     public TaskList() {
         this.list = new ArrayList<>();
+        this.ui = new Ui();
     }
 
     public void addToList(Task task) {
         try {
             this.list.add(task);
-            System.out.println(line);
-            System.out.println("Got it. I've added this task:");
-            System.out.println(task.toString());
-            System.out.println("Now you have " + list.size() + " tasks in the list.");
-            System.out.println(line);
+            ui.addToListMsg(task, this.list.size());
         } catch (Exception e) {
-            System.out.println("Fail to add the task: wrong format");
+            ui.addFailMsg();
         }
     }
 
@@ -26,14 +24,9 @@ public class TaskList {
         if (num <= list.size() && num > 0) {
             Task currTask = list.get(num - 1);
             currTask.markDone();
-            System.out.println(line);
-            System.out.println("Nice! I've marked this task as done:");
-            System.out.println(currTask.toString());
-            System.out.println(line);
+            ui.markMsg(currTask);
         } else {
-            System.out.println(line);
-            System.out.println("Invalid number");
-            System.out.println(line);
+            ui.invalidNumMsg();
         }
     }
 
@@ -41,14 +34,9 @@ public class TaskList {
         if (num <= list.size() && num > 0) {
             Task currTask = list.get(num - 1);
             currTask.markUndone();
-            System.out.println(line);
-            System.out.println("OK, I've marked this task as not done yet:");
-            System.out.println(currTask.toString());
-            System.out.println(line);
+            ui.unMarkMsg(currTask);
         } else {
-            System.out.println(line);
-            System.out.println("Invalid number");
-            System.out.println(line);
+            ui.invalidNumMsg();
         }
     }
 
@@ -56,15 +44,9 @@ public class TaskList {
         if (num <= list.size() && num > 0) {
             Task currTask = list.get(num - 1);
             this.list.remove(currTask);
-            System.out.println(line);
-            System.out.println("Noted. I've removed this task:");
-            System.out.println(currTask.toString());
-            System.out.println(" Now you have " + list.size() + " tasks in the list.");
-            System.out.println(line);
+            ui.deleteMsg(currTask, list.size());
         } else {
-            System.out.println(line);
-            System.out.println("Invalid number");
-            System.out.println(line);
+            ui.invalidNumMsg();
         }
     }
